requirements.txt
dash==1.14.0
dash-core-components==1.10.2
dash-html-components==1.0.3
dash-renderer==1.6.0
dash-table==4.9.0
Flask-Login==0.5.0
Flask-SQLAlchemy==2.4.4
psycopg2==2.8.5
python-dotenv==0.14.0
gunicorn==20.0.4
